move hooks to a helper file
