--- malty/molecules/AlertBanner/hooks/usePrevious.tsx
@@ -1,9 +1,0 @@
-import { useEffect, useRef } from "react";
-
-export const usePrevious = (value: number) => {
-  const ref = useRef<number>(0);
-  useEffect(() => {
-    ref.current = value;
-  }, [value]);
-  return ref.current;
-}
